move non pages out of pages dir
