+++ website/src/pages/swatch/index.astro
@@ -1,5 +1,5 @@
 ---
-import { getMyPatterns } from './list.json';
+import { getMyPatterns } from '../../my_patterns.js';
 
 import { Content } from '../../../../my-patterns/README.md';
 import HeadCommon from '../../components/HeadCommon.astro';
@@ -37,3 +37,4 @@ const baseNoTrailing = BASE_URL.endsWith('/') ? BASE_URL.slice(0, -1) : BASE_URL
     }
   </div>
 </body>
+../../list.json
